Title: 삼 오 무, Time: 7ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240127/삼 오 무/three-five-moo.cpp
@@ -16,7 +16,7 @@ int main() {
 	long long right = INT_MAX;
 	long long dap = INT_MAX;
 	while (left <= right) {
-		int mid = (left + right) / 2;
+		long long mid = (left + right) / 2;
 		if (IsPossible(mid)){
             right = mid - 1;
             dap = min(dap,mid);
